# Fix ECR provider startup latency