### Background
* Before this change, even on non-AWS platforms, the Enabled() check attempts
  to make calls to the metadata endpoint when the session and credentials
  are initialized (in order to determine if the provider should be
  initialized at all).
* This can cause latency because the SDK times out and retries -- up to
  20 seconds of latency has been observed on non-AWS platforms when the
  metadata IP was blocked with an iptables rule.
* Instead, check once if we are running on an EC2 platform, first trying
  to find the EC2 UUID in system files, and second attempting to get
  credentials.
* Add a benchmark test that includes intialization and the credential
  check.

### Benchmark Results

#### GCE, k8s v1.16.1
```
nic@kubernetes-build:~/kubernetes/pkg/credentialprovider/aws$ go test -v -bench=. -run=SetupLatency
goos: linux
goarch: amd64
pkg: k8s.io/kubernetes/pkg/credentialprovider/aws
cpu: Intel(R) Xeon(R) CPU @ 2.30GHz
BenchmarkSetupLatency
E0312 22:29:38.488946   86306 aws_credentials.go:77] while getting AWS credentials NoCredentialProviders: no valid providers in chain. Deprecated.
        For verbose messaging see aws.Config.CredentialsChainVerboseErrors
[...]
BenchmarkSetupLatency-8         1000000000               0.0006955 ns/op
PASS
ok      k8s.io/kubernetes/pkg/credentialprovider/aws    0.032s

```

#### GCE, k8s v1.16.1, with metadata blocked
```
nic@kubernetes-build:~/kubernetes/pkg/credentialprovider/aws$ sudo iptables --insert OUTPUT 1 --destination 169.254.169.254/32 --jump DROP
nic@kubernetes-build:~/kubernetes/pkg/credentialprovider/aws$ go test -v -bench=. -run=SetupLatency
goos: linux
goarch: amd64
pkg: k8s.io/kubernetes/pkg/credentialprovider/aws
cpu: Intel(R) Xeon(R) CPU @ 2.30GHz
BenchmarkSetupLatency
E0312 17:57:50.922473   79323 aws_credentials.go:77] while getting AWS credentials NoCredentialProviders: no valid providers in chain. Deprecated.
        For verbose messaging see aws.Config.CredentialsChainVerboseErrors
BenchmarkSetupLatency-8                1        20308753891 ns/op
PASS
ok      k8s.io/kubernetes/pkg/credentialprovider/aws    20.330s

```

#### GCE, k8s HEAD
```
nic@kubernetes-build:~/kubernetes/pkg/credentialprovider/aws$ go test -v -bench=. -run=SetupLatency
goos: linux
goarch: amd64
pkg: k8s.io/kubernetes/pkg/credentialprovider/aws
cpu: Intel(R) Xeon(R) CPU @ 2.30GHz
BenchmarkSetupLatency
E0312 22:02:59.904691   85420 aws_credentials.go:77] while getting AWS credentials NoCredentialProviders: no valid providers in chain. Deprecated.
        For verbose messaging see aws.Config.CredentialsChainVerboseErrors
[...]
        For verbose messaging see aws.Config.CredentialsChainVerboseErrors
BenchmarkSetupLatency-8         1000000000               0.2403 ns/op
PASS
ok      k8s.io/kubernetes/pkg/credentialprovider/aws    4.542s

```
#### GCE, k8s HEAD, with metadata blocked
```
nic@kubernetes-build:~/kubernetes/pkg/credentialprovider/aws$ sudo iptables --insert OUTPUT 1 --destination 169.254.169.254/32 --jump DROP
nic@kubernetes-build:~/kubernetes/pkg/credentialprovider/aws$ go test -v -bench=. -run=SetupLatency
goos: linux
goarch: amd64
pkg: k8s.io/kubernetes/pkg/credentialprovider/aws
cpu: Intel(R) Xeon(R) CPU @ 2.30GHz
BenchmarkSetupLatency
E0312 22:01:36.837444   85274 aws_credentials.go:77] while getting AWS credentials NoCredentialProviders: no valid providers in chain. Deprecated.
        For verbose messaging see aws.Config.CredentialsChainVerboseErrors
BenchmarkSetupLatency-8                1        6291492407 ns/op
PASS
ok      k8s.io/kubernetes/pkg/credentialprovider/aws    6.309s
```

#### GCE, this commit
```
nic@kubernetes-build:~/kubernetes/pkg/credentialprovider/aws$ go test -v -bench=. -run=SetupLatency
goos: linux
goarch: amd64
pkg: k8s.io/kubernetes/pkg/credentialprovider/aws
cpu: Intel(R) Xeon(R) CPU @ 2.30GHz
BenchmarkSetupLatency
BenchmarkSetupLatency-8         1000000000               0.0000006 ns/op
PASS
ok      k8s.io/kubernetes/pkg/credentialprovider/aws    0.026s
```

#### GCE, this commit, with metadata blocked
```
nic@kubernetes-build:~/kubernetes/pkg/credentialprovider/aws$ sudo iptables --insert OUTPUT 1 --destination 169.254.169.254/32 --jump DROP
nic@kubernetes-build:~/kubernetes/pkg/credentialprovider/aws$ go test -v -bench=. -run=SetupLatency
goos: linux
goarch: amd64
pkg: k8s.io/kubernetes/pkg/credentialprovider/aws
cpu: Intel(R) Xeon(R) CPU @ 2.30GHz
BenchmarkSetupLatency
BenchmarkSetupLatency-8         1000000000               0.0000005 ns/op
PASS
ok      k8s.io/kubernetes/pkg/credentialprovider/aws    0.018s
```

<!--  Thanks for sending a pull request!  Here are some tips for you:

1. If this is your first time, please read our contributor guidelines: https://git.k8s.io/community/contributors/guide/first-contribution.md#your-first-contribution and developer guide https://git.k8s.io/community/contributors/devel/development.md#development-guide
2. Please label this pull request according to what type of issue you are addressing, especially if this is a release targeted pull request. For reference on required PR/issue labels, read here:
https://git.k8s.io/community/contributors/devel/sig-release/release.md#issuepr-kind-label
3. Ensure you have added or ran the appropriate tests for your PR: https://git.k8s.io/community/contributors/devel/sig-testing/testing.md
4. If you want *faster* PR reviews, read how: https://git.k8s.io/community/contributors/guide/pull-requests.md#best-practices-for-faster-reviews
5. If the PR is unfinished, see how to mark it: https://git.k8s.io/community/contributors/guide/pull-requests.md#marking-unfinished-pull-requests
-->

**What type of PR is this?**
> Uncomment only one ` /kind <>` line, hit enter to put that in a new line, and remove leading whitespace from that line:
>

/kind bug

**What this PR does / why we need it**:
Reduce latency introduced by the ECR credential provider, even on non-AWS platforms.

**Which issue(s) this PR fixes**:
<!--
*Automatically closes linked issue when PR is merged.
Usage: `Fixes #<issue number>`, or `Fixes (paste link of issue)`.
_If PR is about `failing-tests or flakes`, please post the related issues/tests in a comment and do not use `Fixes`_*
-->
Fixes #92162

**Special notes for your reviewer**:

**Does this PR introduce a user-facing change?**:
<!--
If no, just write "NONE" in the release-note block below.
If yes, a release note is required:
Enter your extended release note in the block below. If the PR requires additional action from users switching to the new release, include the string "action required".

For more information on release notes see: https://git.k8s.io/community/contributors/guide/release-notes.md
-->
```release-note
Reduces delay initializing on non-AWS platforms docker runtime.
```

**Additional documentation e.g., KEPs (Kubernetes Enhancement Proposals), usage docs, etc.**:

<!--
This section can be blank if this pull request does not require a release note.

When adding links which point to resources within git repositories, like
KEPs or supporting documentation, please reference a specific commit and avoid
linking directly to the master branch. This ensures that links reference a
specific point in time, rather than a document that may change over time.

See here for guidance on getting permanent links to files: https://help.github.com/en/articles/getting-permanent-links-to-files

Please use the following format for linking documentation:
- [KEP]: <link>
- [Usage]: <link>
- [Other doc]: <link>
-->
```docs

```