@therc 2020-07-20
> Need to do some testing.
> /hold

I think this is bringing the code back to my original implementation, so it should work (I checked that it wouldn't introduce delays outside of AWS), unless there's something else that changed substantially in the credential providers. I can approve in the meantime, but please proceed with more testing.