@nckturner 2020-08-31
Did some testing on AWS platforms to make sure this change does not affect image pull for both credentialed (ECR) and non-credentialed pulls, and it looks fine.