@andrewsykim 2020-11-18
On second thought, this could be misleading because kubelet will always attempt to authenticate to ECR repos but `Provide` will fail if there's no session established. Could `Enabled` instead check the availablility of the metadata server and/or the aws credentials on disk `~/.aws/credentials`?