@nckturner 2020-11-18
Will address the comments