@nckturner 2020-11-18
@andrewsykim 
> On second thought, this could be misleading because kubelet will always attempt to authenticate to ECR repos but `Provide` will fail if there's no session established. Could `Enabled` instead check the availablility of the metadata server and/or the aws credentials on disk `~/.aws/credentials`? Basically a similar check to what already exists but skips the API calls.

Hmm, I'm not sure what assumptions we want to make here.  Metadata isn't guaranteed to be available, and in the cases where its not, we'd need to check everything in the sdk's credential chain (i.e., one possibility is using the 2 env vars: `AWS_WEB_IDENTITY_TOKEN_FILE` and `AWS_ROLE_ARN`).  Probably possible but also sounds brittle.