@andrewsykim 2020-11-18
> we'd need to check everything in the sdk's credential chain

Is there a way to configure the client to check everything in the SDK credential chain except for those that require API calls?