@nckturner 2020-11-19
@andrewsykim I actually think that its retries to IMDS during the normal that is causing the latency when we try to set up the session. There is logic in the SDK which has to decide between IMDSv1 and IMDSv2, with a fallback mechanism.  There have been issues in the past with the SDK which caused latency when falling back to IMDSv1.  If this is the case, then we might be able to get rid of the latency just adjusting the timeout.  The other possibility is, the SDK sees the service account token and then tries to do an assume-role-with-web-identity call, which it can use to exchange for aws creds.  I'll see if I can pinpoint the latency.