@nckturner 2020-11-20
@andrewsykim I repro'd this on a gce instance (although that might have not been necessary, given the repro).  If you drop packets to the metadata endpoint via an iptables rule, the aws sdk gets really unhappy:  

```
root@kubernetes-build:/home/nic/kubernetes/pkg/credentialprovider/aws# sudo iptables -A OUTPUT -p tcp -o ens4 -d 169.254.169.254/32 -j DROP
root@kubernetes-build:/home/nic/kubernetes/pkg/credentialprovider/aws# telnet 169.254.169.254 80
Trying 169.254.169.254...
^C
root@kubernetes-build:/home/nic/kubernetes/pkg/credentialprovider/aws# go test -bench .
E1120 07:51:03.065014    9083 aws_credentials.go:69] aws_credential:1
E1120 07:51:03.065656    9083 aws_credentials.go:73] aws_credential:2
E1120 07:51:03.065686    9083 aws_credentials.go:78] aws_credential:3
E1120 07:51:23.405789    9083 aws_credentials.go:80] aws_credential:4
E1120 07:51:23.405826    9083 aws_credentials.go:81] while getting AWS credentials NoCredentialProviders: no valid providers in chain. Deprecated.
        For verbose messaging see aws.Config.CredentialsChainVerboseErrors
ECR is enabled: false
goos: linux
goarch: amd64
pkg: k8s.io/kubernetes/pkg/credentialprovider/aws
BenchmarkSetupLatency-8                1        20340895511 ns/op
PASS
ok      k8s.io/kubernetes/pkg/credentialprovider/aws    20.358s
root@kubernetes-build:/home/nic/kubernetes/pkg/credentialprovider/aws# 

```

This is the patch I used to test: 
```
nic@kubernetes-build:~/kubernetes/pkg/credentialprovider/aws$ git diff
diff --git a/pkg/credentialprovider/aws/aws_credentials.go b/pkg/credentialprovider/aws/aws_credentials.go
index a650a7da50b..3a06d9ea7b3 100644
--- a/pkg/credentialprovider/aws/aws_credentials.go
+++ b/pkg/credentialprovider/aws/aws_credentials.go
@@ -66,17 +66,22 @@ func newECRProvider(getterFactory tokenGetterFactory) *ecrProvider {
 // Enabled implements DockerConfigProvider.Enabled. Enabled is true if AWS
 // credentials are found.
 func (p *ecrProvider) Enabled() bool {
+       klog.Error("aws_credential:1")
        sess, err := session.NewSessionWithOptions(session.Options{
                SharedConfigState: session.SharedConfigEnable,
        })
+       klog.Error("aws_credential:2")
        if err != nil {
                klog.Errorf("while validating AWS credentials %v", err)
                return false
        }
+       klog.Error("aws_credential:3")
        if _, err := sess.Config.Credentials.Get(); err != nil {
+               klog.Error("aws_credential:4")
                klog.Errorf("while getting AWS credentials %v", err)
                return false
        }
+       klog.Error("aws_credential:5")
        return true
 }
 
diff --git a/pkg/credentialprovider/aws/aws_credentials_test.go b/pkg/credentialprovider/aws/aws_credentials_test.go
index 7161fc136d1..27536acce20 100644
--- a/pkg/credentialprovider/aws/aws_credentials_test.go
+++ b/pkg/credentialprovider/aws/aws_credentials_test.go
@@ -331,3 +331,9 @@ func TestChinaECRProvideCached(t *testing.T) {
                t.Errorf("cached credentials do not match")
        }
 }
+
+func BenchmarkSetupLatency(b *testing.B) {
+       p := newECRProvider(&ecrTokenGetterFactory{cache: make(map[string]tokenGetter)})
+       enabled := p.Enabled()
+       fmt.Printf("ECR is enabled: %v\n", enabled)
+}

```

Without the iptables rule, the benchmark takes about 30 milliseconds:
```
ok      k8s.io/kubernetes/pkg/credentialprovider/aws    0.031s
```
and from a network namespace without any connectivity: 
```
root@kubernetes-build:/home/nic/kubernetes/pkg/credentialprovider/aws# ip netns exec test0 go test -bench .
...
ok      k8s.io/kubernetes/pkg/credentialprovider/aws    0.021s

```
Its about 10 ms faster then the default case, probably because it immediately fails because there is no interface. (20 ms is still a long time though...)

Anyway, this is all to say, I don't think we can rely on the SDK or configure it to only check "the fast things" since instance metadata is pretty central to what it tries to do.  Using `-j REJECT` instead of `-j DROP brings the latency down to 4-6 seconds (depending on version of sdk), but that's still unacceptable. I think I'm leaning toward's @liggitt's solution of a run once check, the only thing I don't like about it is that if the metadata check randomly fails for an AWS kubelet, then that kubelet would have to be restarted (if I'm understanding correctly).