@andrewsykim 2020-11-20
> the only thing I don't like about it is that if the metadata check randomly fails for an AWS kubelet, then that kubelet would have to be restarted (if I'm understanding correctly).

Yeah this would be problematic -- a failed connection to metadata server would disable the provider, but we can't necessary panic either,  otherwise kubelets not on EC2 instances would panic too.