@liggitt 2020-11-20
> the only thing I don't like about it is that if the metadata check randomly fails for an AWS kubelet, then that kubelet would have to be restarted (if I'm understanding correctly).

it would be nice if that could be improved, but I see that as something for the AWS provider to work on as a follow-up after getting this merged to avoid affecting all non-AWS deployments